feat: add type definition, export symbols
This makes it possible to use this library in a typescript environment,
and take advantage of the indent/newline detection.

Without exporting the symbols themselves, typescript gets confused about
the "unique symbol", since it isn't clever enough to verify that the
same string literal passed to Symbol.for will result in the same Symbol
object.  But, it does track object identities, so exposing the symbols
on the export works fine.

Alternatively, exporting a `parseJson.stringify` method would be even
handier, since then the consumer wouldn't have to worry about pulling
the indent/newline off of the object.